Move hostcalls impl into separate module
