+++ src/hostcalls_impl/mod.rs
@@ -1,0 +1,5 @@
+mod fs;
+mod misc;
+
+pub(crate) use self::fs::*;
+pub(crate) use self::misc::*;
